Added support for includes of inherited elements in inherited composers.
This involved a major rework of how the response is written back to the client. It is not buffered locally until everything has been written to it, then the headers are added and only then is the response being written back to the client.

Note that this commit breaks functionality with the cache.