[Feature] Add information about origin & original identifier to resource
This commit introduces to new fields on a resource object: the origin (name)
and the original identifier. With the two new properties, a resource, which
originates from an other system, can now be tracked by it's origin and the
original identifier.

As a part of this change, the previously known property 'source' of the
ResourceContent interface got removed. This is a breaking API change!