[Bugfix] Fix problem with request protocol miss-match
The JAXRSServlet checks if a request maps to the site which
registered the requested endpoint. This check was done by
comparing the full URLs including their protocol as well as the
port.

This behavior lead to problems if the protocol/port specified
in the site's URLs (site.xml) did not match the protocol/port of the
request.

This fix changes the behavior to only check if the requested
hostname matches a hostname in the site's URLs and does not
take the protocol nor the port into account anymore.