ENTV-229 [Improvement] Rename resource field 'type' to 'flavor'
Rename the existing resource field 'type' to 'flavor' to differentiate between
the the resource type like Page, Image, Movie or File and the flavor of such a
resource like 'News', 'Episode' or 'Series'

This change breaks backward compatibility, since the XML structure of a serial-
ized resource has changed! Only use this with a clean repository/index.