[Improvement] Replaced the Security Service with the static SecurityUtils to obtain a thread's current user and site
This change simplifies dealing with a thread's security information such as the current user and site and eliminates the need to pass around the security service that used to hold this information.

While the SecurityService api has been removed, the former SpringSecurityService has been renamed to SecurityServiceImpl and is now used to install implementations of java.security.SecurityManager as the System's security manager.

This change requires adapted (simplified) versions of the spring security configuration. Basically, all mentions of the SecurityService need to be removed (see sample security.xml as part of this commit).

+Configuration

Conflicts:
	modules/weblounge-common-api/src/main/java/ch/entwine/weblounge/common/site/Site.java