ENTV-245 [Improvement] Add access information to index
Add support for storing the access information in the resource index. This
allows to build queries that only return those resource that the current
user is allowed to perform the given action (read, write, ...) on.

Important note:
This change requires updating the index configuration as well as rebuilding
the index!

+Configuration
+Index