Issue #373 - Remove module 'bridge'
The original purpose of the Weblounge bridge module was to connect third
party systems to serve as content producers.

The mail bridge never made it to production quality and only served as
proof of concept. Even though the feature itself sounds interesting, the
existing code is of no real value and should be removed.

The OAI-PMH harvester (copied from the Opencast project, see
http://opencast.org) was used to as base class in a closed-source
project. As this is the only known implementation and it not directly
relates to Weblounge, it should be safe to remove that part of the
module as well.